refactor: Rename `lib/plugins/aws/package/compile/events/cloudFront.js`
