+++ lib/plugins/aws/package/compile/events/cloud-front.js
