+++ lib/plugins/index.js
@@ -51,7 +51,7 @@ module.exports = [
   require('./aws/package/compile/events/cognitoUserPool.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs.js'),
-  require('./aws/package/compile/events/cloudFront.js'),
+  require('./aws/package/compile/events/cloud-front.js'),
   require('./aws/package/compile/events/httpApi'),
   require('./aws/deploy-function.js'),
   require('./aws/deploy-list.js'),
